Initial implementation of async emitting based on LatencyQueue. Original intended functionality was naive on at least three counts: First, silently dropping messages on the floor may not be desirable. Second, it may take multiple consumer threads to drain one producer thread. Third: There should be some way to wait for the queue to flush. Details:
Silently dropping messages: A static capacity queue likely remains the right choice to avoid buffer-bloat/limit the amount of pending debt--since the producer only shoulders ~10% of the cost of logging to a file descriptor, it's easy for producers to outstrip consumers. Dropping messages isn't neccissarily wrong in that situation; however, behavior should be configurable rather than mandated. Waiting isn't without complications either: forever isn't ideal either. Likely we want a wait_for() style behavior, and allow the end-user to choose how long. LatencyQueue will need to be updated to reflect that: IngressNode will need a condition variable and a way to register intrest for a wakeup when no empty caches are available; ReceiveNode will need to check for and wake waiting IngressNode when it makes empty cahces available. LatencyQueue's insert signature should accept a duration for waiting as an argument--an empty duration is equivalent to current insert behavior.

Draining the backlog(producer/consumer asymmetry): Creating a message object is likely less than 10% of the work of emitting it. However, LatencyQueue currently doesn't accomodate multiple consumers per producer. LatencyQueue was designed to be low contention under moderate load: A single consumer would service multiple producers; producers buffer output, offloading and waking the consumer only when the buffer is full; when the consumer isn't busy, it steals buffered work from producers at a regular interval to ensure a maximum latency. However, when a producer is saturating the queue, the consumer is driven by push behavior and doesn't scan other producers. That's fine unless the consumer never catches up, which can be the case here with a noisy producer. Right now LatencyQueue only offers one mode of consumer parallelism: strictly partitioning producers among consumers. Since strict partitioning has good contention characteristics at scale, it makes sense to preserve that and extend LatencyQueue to accomodate multiple consumers per partition. This is a little controversial: With multiple consumers potentially draining one producer, out-of-order emitting from the same thread will happen(Right now LatencyQueue promises in-order from the same producer).

As far as implementing multiple consumers per producer, a good candidate would be to modularize and replicate the de-queuing state in ReceiveNode: Each ReaderNode would contain a 'drain' queue, 'cache' and 'item' iterators, and 'occupied' state. Consumers would continue to share ReceiveNode's mutex and condition variable, so that producers only wake one consumer at a time in each partition. For simplicity, the number of ReaderNodes per ReceiveNode can be an initialization-time decision applied to all nodes--async::Dispatcher can be replaced when reconfigured. With a constant number of read-heads, the signature for remove would have have to accept a ReaderNode index in addition to a ReceiverNode index. While slightly more cumbersome, it obviates the need to handle partially read queue contents when removing a ReaderNode.

Flushing the queue: it's impossible to actually gauge the performance of the async implementation without a flush() mechanic on dispatcher. Since the number of IngressNodes is static, that could be done by creating a Message::Contents implementation that appropriated stream() to count invocations. LatencyQueue's insert would have to be expanded to accept an ingress index for that to work. In all, Dispatcher::flush() would create and return a flush type with atomic count and condition variable, and inqueue one message per ingress that operate those variables when streamed. The end user would then wait for the condition to be notified.

In all, lots of behavioral expansions appear to be in order.